Fix Message Displayer
