+++ src/components/chat-item/Chat.module.scss
@@ -20,14 +20,6 @@
     box-shadow: none;
   }
 
-  .avatar {
-    @extend .flexCenter;
-    width: 50px;
-    height: 50px;
-    border-radius: 1000px;
-    overflow: hidden;
-  }
-
   .left {
     @extend .justifyBtw;
     flex: 1;
