+++ src/components/message-displayer/MsgDisplayer.tsx
@@ -1,47 +1,68 @@
 import React, { useEffect } from "react";
 import clsx from "clsx";
+import {
+  collection,
+  doc,
+  onSnapshot,
+  orderBy,
+  query,
+} from "firebase/firestore";
 
+import { db } from "@/pages/api/firebase";
 import { formatDate } from "src/lib";
 import EnterMsg from "components/enter-msg";
 import Spinner from "components/spinner";
 import Message from "components/message/Message";
+import EmptyScreen from "components/empty-screen";
 
 import stl from "./MsgDisplayer.module.scss";
 
 interface Props {
-  messages: any;
   myId: string;
   theme: string;
   chatId: string;
-  isLoading: Boolean;
-  setIsLoading: any;
+  setTitle: (arg: any) => void;
+  isEmpty: Boolean;
 }
 
-const MsgDisplayer = ({
-  messages,
-  myId,
-  theme,
-  chatId,
-  isLoading,
-  setIsLoading,
-}: Props) => {
-  const ref = React.useRef(null);
+const MsgDisplayer = ({ myId, theme, chatId, setTitle, isEmpty }: Props) => {
+  const [isLoading, setIsLoading] = React.useState(false);
+  const [messages, setMessages] = React.useState([]);
 
   useEffect(() => {
-    setTimeout(() => setIsLoading(false), 1000);
-  }, []);
+    setIsLoading(true);
+    const chatsRef = collection(db, "chats");
+    const chatDoc = doc(chatsRef, chatId);
+    onSnapshot(chatDoc, (snapshot) => setTitle(snapshot.data()?.chatName));
+    const msgRef = collection(chatDoc, "messages");
+    const sortedMsgs = query(msgRef, orderBy("time", "asc"));
+    onSnapshot(sortedMsgs, (snapshot) => {
+      //@ts-ignore
+      let msgsArray = [];
+      snapshot.docs.map((doc, i) =>
+        msgsArray.push({ ...doc.data(), id: doc.id, key: i })
+      );
+      //@ts-ignore
+      setMessages(msgsArray);
+      setIsLoading(false);
+    });
+  }, [chatId]);
+
+  const ref = React.useRef(null);
 
   useEffect(() => {
     setTimeout(() => {
       // @ts-ignore
       ref.current.scrollTop = ref.current.scrollHeight;
-    }, 1500);
+    }, 2000);
   }, []);
 
   return (
     <div className={stl.msgDisplayer}>
       <div ref={ref} className={stl.msgContainer}>
-        {isLoading ? (
+        {isEmpty ? (
+          <EmptyScreen />
+        ) : isLoading ? (
           <div className={stl.loadingContainer}>
             <Spinner />
           </div>
