+++ src/components/verify-msg/VerifyMsg.module.scss
@@ -1,6 +1,6 @@
 .verifyMsg {
   width: 100%;
-  height: 100%;
+  height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
